ete_educa_v3_full/requirements.txt
streamlit>=1.37.0
openai>=1.51.0
python-dotenv>=1.0.1
